authorization basic auth yerine jwt token ile sağlandı

+++ src/main/java/com/workintech/twitter/service/RetweetServiceImpl.java
@@ -4,14 +4,18 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import com.workintech.twitter.dto.request.RetweetRequestDto;
 import com.workintech.twitter.dto.response.RetweetResponseDto;
 import com.workintech.twitter.entity.Retweet;
 import com.workintech.twitter.exceptions.RetweetNotFoundException;
+import com.workintech.twitter.exceptions.UnauthorizedDeleteException;
 import com.workintech.twitter.mapper.RetweetMapper;
 import com.workintech.twitter.repository.RetweetRepository;
+import com.workintech.twitter.repository.TweetRepository;
+import com.workintech.twitter.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,6 +28,12 @@ public class RetweetServiceImpl implements RetweetService {
 
     @Autowired
     private final RetweetMapper retweetMapper;
+
+    @Autowired
+    private final UserRepository userRepository;
+
+    @Autowired
+    private final TweetRepository tweetRepository;
 
     @Override
     public List<RetweetResponseDto> getAll() {
@@ -46,13 +56,41 @@ public class RetweetServiceImpl implements RetweetService {
 
     @Override
     public RetweetResponseDto create(RetweetRequestDto retweetRequestDto) {
-         Retweet retweet = retweetMapper.toEntity(retweetRequestDto);
-       retweet = retweetRepository.save(retweet);
-       return retweetMapper.toRetweetResponseDto(retweet); 
+        org.springframework.security.core.Authentication authentication = org.springframework.security.core.context.SecurityContextHolder.getContext().getAuthentication();
+        String email = authentication.getName();
+        com.workintech.twitter.entity.User user = userRepository.findByEmail(email)
+            .orElseThrow(() -> new com.workintech.twitter.exceptions.UserNotFoundException("Kullanıcı bulunamadı"));
+
+        com.workintech.twitter.entity.Tweet tweet = tweetRepository.findById(retweetRequestDto.tweetId())
+            .orElseThrow(() -> new com.workintech.twitter.exceptions.TweetNotFoundException("Tweet bulunamadı"));
+
+        java.util.Optional<Retweet> existingRetweet = retweetRepository.findByUserIdAndTweetId(user.getId(), tweet.getId());
+        if (existingRetweet.isPresent()) {
+            retweetRepository.delete(existingRetweet.get());
+            return null; // veya özel bir response dönebilirsiniz
+        } else {
+            Retweet retweet = retweetMapper.toEntity(retweetRequestDto);
+            retweet.setUser(user);
+            retweet.setTweet(tweet);
+            retweet = retweetRepository.save(retweet);
+            return retweetMapper.toRetweetResponseDto(retweet);
+        }
     }
     
     @Override
     public void deleteById(Long id) {
+        org.springframework.security.core.Authentication authentication = org.springframework.security.core.context.SecurityContextHolder.getContext().getAuthentication();
+        String email = authentication.getName();
+        com.workintech.twitter.entity.User user = userRepository.findByEmail(email)
+          .orElseThrow(() -> new com.workintech.twitter.exceptions.UserNotFoundException("Kullanıcı bulunamadı"));
+
+        Retweet retweet = retweetRepository.findById(id)
+          .orElseThrow(() -> new com.workintech.twitter.exceptions.RetweetNotFoundException(id + " id'li retweet bulunamadı."));
+
+        if (retweet.getUser() == null || !java.util.Objects.equals(user.getId(), retweet.getUser().getId())) {
+          throw new UnauthorizedDeleteException("Bu retweet'i silmeye yetkiniz yok.", HttpStatus.FORBIDDEN);
+        }
+
         retweetRepository.deleteById(id);
     }
